Bug fix: install/activation routine would never run
WordPress hooks are callbacks. Callback can only call `public` methods in a class when they are called from outside of the class - as is the case for hook -.

In other words, the `activate()` method would never run and would throw a PHP warning `'call_user_func_array() expects parameter 1 to be a valid callback, cannot access private method ZeroSpam_Install::activate()`, which would go unnoticed in most cases as in PHP errors are normally hidden in active sites.

Fixes:
* https://wordpress.org/support/topic/activation-error-78/
* https://wordpress.org/support/topic/plugin-is-generating-errors/
* https://wordpress.org/support/topic/plugin-activation-error-17/