Add Composer config file
This will allow the plugin to be hosted on [Packagist](http://packagist.org/) and for it to be installed via the PHP dependency manager [Composer](https://getcomposer.org/).

For now, the minimum PHP requirement has been set to PHP 5.2.

Once support for WP < 5.2 has been dropped, this can be upped to PHP 5.6.