Add a configuration for Travis CI
This adds a minimal configuration for Travis CI which, for now, just does a basic QA check:
* It checks that the PHP code does not contain any parse errors on select high/low PHP versions for each PHP major.

For this to work, Travis CI needs to be turned on for this repo. (free for open source)
See this tutorial: https://docs.travis-ci.com/user/tutorial/

Once this PR has been merged and Travis CI activated for this repo, the script will run on each PR to prevent parse errors from entering the repo.