Simplify the uninstall file
The `uninstall.php` file contained a lot of redundant checks.

WordPress will not call this file _unless_ uninstalling of this plugin was requested. As a plugin you don't need to duplicate the checks WP already does before calling the file.

The only relevant checks are that the user has the right to uninstall and that the file is called by WordPress while uninstalling plugins.

Similarly, there is no need to check whether options exist before deleting them. If they don't, WP will just return `false` and do nothing, so no harm done.

This incidentally also gets rid of a PHP cross-version compatibility issue as the `__DIR__` magic constant was used in the code now removed and `__DIR__` was only introduced in PHP 5.3, so the uninstall would fail when run on PHP 5.2.