Bug fix: option may not be set
As the install routine will not have run for any of the installs in the past two and half years, the default settings may not have been set.

For this particular setting, this was causing an `Undefined index: blocked_ip_msg` PHP notice to be thrown.

This minor change fixes that.

Fixes 147

Fixes:
* https://wordpress.org/support/topic/notice-undefined-index-blocked_ip_msg/