Bug fix: fix parse error
PHP short open tags are only supported when the PHP ini directive `short_open_tag` is set to `true` or if PHP was configured with the `--enable-short-tags` option.

Using them is discouraged as it will cause PHP to be displayed inline as if it were HTML when support for short open tags is turned off, which could be considered a security risk.

Another reason for it to be discouraged is that it conflicts with XML declarations.

For that reason on a lot of webservers, the setting is turned off by default, so for cross-server compatible code, full PHP open tags need to be used.

Refs:
* https://www.php.net/manual/en/language.basic-syntax.phptags.php
* https://www.php.net/manual/en/ini.core.php#ini.short-open-tag