QA/performance: remove duplicate function calls
[Part of CS fixing PR series]

A function call in the "condition" part of a `for` loop will be executed for every loop made.

This means in effect that `date( 'Y' )` in this comparison `$i <= ( $zerospam_year + 50 )` is called 51 times while running through the `for`.

As the result shouldn't change during the `for` loop, making the function call once at the start of the script is sufficient.