QA: adjust security check
When a file will directly generate output when called, the "security check" could be considered valid, though a case can be made that if the file does not make use of any of the request variables, such as `$_GET` or `$_POST` and doesn't do any updates (to the database etc), there is no real harm in the plain HTML skeleton being displayed either.

So, based on that, these checks _could_ be removed.

If they remain, however, let's just fail silently without any indication of why the request failed.