QA: don't even risk overriding a WP global variable
The variable `$type` is in use by WP itself.

While this file within the context of this plugin is included from within a function, making all variables declared in it local to that function, there is no _guarantee_ that the file will always be included in this way, so let's just make sure there is no risk whatsoever that the WP global variable `$type` could be overridden by the variable declared in this template file.